+++ pkg/app/image_describe.go
@@ -42,7 +42,7 @@ func getImageDependency(clicontext *cli.Context) error {
 	if envName == "" {
 		return errors.New("image is required")
 	}
-	envdEngine, err := envd.New(clicontext.Context)
+	envdEngine, err := envd.New(clicontext.Context, "docker")
 	if err != nil {
 		return errors.Wrap(err, "failed to create envd engine")
 	}