+++ pkg/lang/ir/interface.go
@@ -66,7 +66,7 @@ func GPU(numGPUs int) {
 
 func CUDA(version, cudnn string) {
 	DefaultGraph.CUDA = &version
-	DefaultGraph.CUDNN = &cudnn
+	DefaultGraph.CUDNN = cudnn
 }
 
 func VSCodePlugins(plugins []string) error {
@@ -112,6 +112,7 @@ func JuliaPackageServer(url string) error {
 
 func Shell(shell string) error {
 	DefaultGraph.Shell = shell
+	DefaultGraph.SystemPackages = append(DefaultGraph.SystemPackages, shell)
 	return nil
 }